fix(solidstart): Use production server for e2e tests (#14033)
Adds a small post-build script to copy over deps that are not correctly
resolved in our yarn workspace/pnpm e2e test applications (but do work
correctly in standalone projects). This allows us to finally use the
vinxi production server to run e2e tests and is similar to what the Nuxt
SDK has to do.

Also allows us to get rid of the double page loads for the first test to
avoid hydration errors that were caused by the dev server previously.