debug(replay): Remove noisy log that is causing event buffer issues as well
Removes a log statement that is output when buffer is already full (which will keep overflowing the buffer). Also noticed a log statement in `stop()` that would trigger when we stop due to event buffer being full, which would compoound the issue. This last fix should be pulled into the SDK.