ref(nextjs): Make build-time value injection turbopack compatible (#14081)
Ref: https://github.com/getsentry/sentry-javascript/issues/8105

To inject build-time variables, in addition to doing so via a custom
loader, we will be injecting them via the `env` option.

Caveat: We are currently using the Next.js build ID as a release name.
This build id is passed to the `webpack` option. Since the `webpack`
option doesn't exist for turbopack we don't have access to the build ID.
For now we will simply not inject a release name, which may be better
anyhow since turbopack is currently only stable for dev.