feat(utils): Single implementation to fetch debug ids (#14199)
In the process of fixing
https://github.com/getsentry/sentry-electron/issues/1011 I was looking
for a way to get a `Map<Filename, DebugId>`. I found that there was
already some code duplication for parsing and caching of debug ids from
the polyfilled globals.

This PR moves the common code to `@sentry/utils` which will be useful
for fixing the above.