feat(v7/browser): Add moduleMetadataIntegration lazy loading support (#13822)
This PR fixes #13803. The corresponding PR for latest major version is
#13817.

I saw the bundles generation for integration is happening in
`@sentry/integrations`, so I added a corresponding file for
`modulemetadata` in this package that exports that integration from
`@sentry/core`, so that the bundle actually gets created for this
integration as needed.