fix(nuxt): Fix dynamic import rollup plugin to work with latest nitro (#14243)
With nitro 2.10.X the way we imported the server config file and
dynamically imported the nitro runtime broke. Instead of landing in the
server output entry file, the sdk init and dynamic import landed in the
nitro runtime itself.

Based on
https://github.com/nitrojs/nitro/commit/8b4a408231bdc222569a32ce109796a41eac4aa6#diff-e58102d2230f95ddeef2662957b48d847a6e891e354cfd0ae6e2e03ce848d1a2R142
we managed to get back to the original solution by using the `\0raw`
prefix for the entry file.

__Note for internal__: This does not play well when using symlinks (like
`yarn link`) to link the sdk into a project. Haven't tested how yalc
behaves, but tarballs work.