chore(dev-deps): Bump nestjs deps & add nestjs v8 E2E test (#14148)
We support v8 but only test v10. By adding a (very basic) e2e test app
for v8 of nestjs we can ensure this continues to work (as long as we
want to support it).

Also bumps nestjs dev deps to hopefully fix some security warnings for
outdates deps (e.g.
https://github.com/getsentry/sentry-javascript/security/dependabot/376).

Note for the future: There seems to be some problem with the
node-integration-tests and module resolution. As soon as the nestjs
version there matches the one from the nestjs package - which means it
will hoist the dependency to the workspace-level node_modules folder -
the tests start failing weirdly. For now, I "fixed" this by using a
different version of nestjs in node-integration-tests from nestjs. Since
we want to get rid of the node-specific part there anyhow in v9, I
figured it's not worth it to put more work into fixing this properly...