fix(core): Set `sentry.source` attribute to `custom` when calling `span.updateName` on `SentrySpan` (#14251)
Fix a "regression" introduced in v8 where we no longer updated
the source of the span when calling `span.updateName`. We had this
behaviour in our v7 `Transaction` class (IIRC via
`transaction.setName(name, source)` but most likely forgot to port it to
the `Span` class as the `event.transaction_info.source` field is only
relevant for transactions (i.e. root spans in today's SDK lingo).