feat(replay): Clear event buffer when full and in buffer mode
This makes a change so that when the event buffer is full, we will clear it and wait for the next checkout instead of erroring and stopping the replay buffering. This does mean that it's possible an exception occurs in-between the filled buffer and the next checkout, where we will have an empty replay.