feat(core): Ensure `normalizedRequest` on `sdkProcessingMetadata` is merged (#14315)
We keep `normalizedRequest` with data like headers, urls etc. on the
`sdkProcessingMetadata` on the scope.

While this generally works, there are some _potential_ pitfalls/footguns
there:

One, if you put some (e.g. partial) `normalizedRequest` on the _current_
scope, it will just overwrite the full `normalizedRequest` from the
_isolation_ scope, where generally we'll try to put the request data
automatically (e.g. in the http instrumentation). Think this example:

```js
Sentry.withScope(scope => {
  scope.setSdkProcessingMetadata({ normalizedRequest: { headers: moreSpecificHeaders } });
});
```

the resulting event inside of this `withScope` callback would _only_
have the headers, but would miss e.g. url etc. data set.

This PR changes this so that the `normalizedRequest` is merged between
the types of scopes, so only set fields on e.g. the current scope will
overwrite the same fields on the isolation scope, instead of just
overwriting the whole normalizedRequest that results.

Note that bundle size for browser is sadly anyhow affected (no matter if
we go with a/b/c), as the code to merge it between scopes is always
shared :(