feat(browser): Send additional LCP timing info (#14372)
Adds `loadTime` and `renderTime` attributes from the [LCP
performance entry](https://developer.mozilla.org/en-US/docs/Web/API/LargestContentfulPaint#instance_properties)
to the `pageload` span. Besides being valuable information for users,
this also helps us detect if LCP is potentially unreliable: If the LCP element was from a 3rd party origin and the
`Timing-Allow-Origin` header was not sent in the resource response, the
`renderTime` will be `0`.