ref(types): Avoid some `any` type casting around `wrap` code (#14463)
We use a heavy dose of `any`, that I would like to reduce.

So I started out to look into `wrap()`, which made heave use of this,
and tried to rewrite it to avoid `any` as much as possible. This
required some changes around it, but should now have much better type
inferrence etc. than before, and be more "realistic" in what it tells
you.

While at this, I also removed the `before` argument that we were not
using anymore - `wrap` is not exported anymore, so this is purely
internal.