fix(opentelemetry): Ignore invalid parent spans (#14512)
We do not align with how OTEL traces and the default samplers handle
invalid span contexts today.
If a span context is invalid, e.g. the span ID or trace ID are invalid,
then the span is ignored for new spans as a parent, and the new span
will start a new trace. This change aligns our own sampling strategy
with this, to ensure that invalid parent span contexts are ignored.