chore: Move `pnpm` version to root package.json & cleanup `clean` command (#14531)
I noticed that running `yarn clean` from the root hang for me sometimes,
in two ways.

On one hand, it seemed to hang/fail on running `pnpm store prune` in the
e2e-tests package. This PR fixes this by moving the pnpm dependency to
the root of the monorepo, ensuring we have a consistent version for this
and volta can properly set this up.

On the other hand, it seemed to hang on watchman cleanup in
node-integration-tests. I moved this up actually, and only run it in
`clean:all` as we don't really need this all that often.

Finally, I also aligned the used rimraf version across the monorepo to
5.0.10.