chore: Dedupe `@sentry/core` imports (#14529)
I noticed we have a whole bunch of files that have multiple imports from
`@sentry/core`, so I went over all of those I could find and ran VSCode
auto-import-ordering to clean this up. Would be nice to have this taken
care of by eslint/biome, but somehow this does not seem to work... :(

This for sure does not capture everything, but a good amount of
duplication at least.