feat(opentelemetry): Stop looking at propagation context for span creation (#14481)
This PR changes the behavior of the OTEL-based Node SDK to ignore the
propagation context when starting spans.

Previously, when you called `startSpan` and there was no incoming trace,
we would ensure that the new span has the trace ID + span ID from the
propagation context.

This has a few problems:

1. Multiple parallel root spans will continue the same virtual trace,
instead of having separate traces.
2. This is really invalid in OTEL, as we have to provide a span ID and
cannot really tell it to use a specific trace ID out of the box. Because
of this, we had to add a bunch of special handling to ensure we can
differentiate real and fake parent span IDs properly.

This PR fixes this by simply not looking at this anymore. For TWP and
error marking the propagation context is still used as before, only for
new spans is there a difference.

I also added docs explaining how trace propagation in node works now:


![node-sdk-trace-propagation-3](https://github.com/user-attachments/assets/73cc5972-2d2a-438c-9c32-2551fc52568e)