feat(core): Add `handled` option to `captureConsoleIntegration` (#14664)
Add a `handled` option to the `captureConsoleIntegration`.
Right now, for v8, it will default to `false` as previously but it gives
users the option to override the handled value. In v9, we will flip the
default value to `true` after discussing this offline with the team. In
this integration we can never be sure about the `handled` state at all.
But flipping it to handled by default seems reasonable.

Co-authored-by: Luca Forstner <luca.forstner@sentry.io>