ref(browser): Update `supportsHistory` check & history usage (#14696)
We had some elaborate check in there for `window.history` which is based
on very old (~2016) issues, which should not be an issue anymore in
versions we support today. So we can streamline this quite a bit!

Additionally, I also rewrote code that used `window.onpushstate` to
instead of `window.addEventListener('pushstate')`, plus also added code
to ensure we do not emit history changes multiple times, which may be
possible (?).