ref(core)!: Remove `Scope` type interface in favor of using `Scope` class (#14721)
In v8, types have been exported from `@sentry/types`, while
implementations have been exported from other classes.
This lead to some duplication, where we had to keep an interface in
`@sentry/types`, while the implementation mirroring that interface was
kept e.g. in `@sentry/core`.
Since in v9 the types have been merged into `@sentry/core`, we can get
rid of some of this duplication. This means that certain things that
used to be a separate interface, will not expect an actual instance of
the class/concrete implementation. This should not affect most users,
unless you relied on passing things with a similar shape to internal
methods.

This PR removes the `Scope` interface, in favor of just using the scope
class everywhere.

This is related to
https://github.com/getsentry/sentry-javascript/issues/9840

---------

Co-authored-by: Sigrid Huemer <32902192+s1gr1d@users.noreply.github.com>