feat(v8/node/deps): Bump @prisma/instrumentation from 5.19.1 to 5.22.0 (#14755)
Backport of https://github.com/getsentry/sentry-javascript/pull/14753:

Bump
[@prisma/instrumentation](https://github.com/prisma/prisma/tree/HEAD/packages/instrumentation)
from 5.19.1 to 5.22.0.

See https://github.com/prisma/prisma/releases/tag/5.22.0 for release
notes.

This is to take advantage of the following change in 5.22.0:

> In our ongoing effort to stabilize the tracing Preview feature, we’ve
made our spans compliant with OpenTelemetry Semantic Conventions for
Database Client Calls. **This should lead to better compatibility with
tools such as DataDog and Sentry** [emphasis added by me].

There's also this change in
[5.21.0](https://github.com/prisma/prisma/releases/tag/5.21.0):

> The `tracing` Preview feature now has full support for MongoDB with
previously missing functionality now implemented. This is a part of the
ongoing effort to stabilize this Preview feature and release it in
General Availability.

The @dependabot PR for `@prisma/instrumentation` updates this to 6.0.0,
which unfortunately drops Node.JS v18 support, so it isn't suitable for
Sentry v8: https://github.com/getsentry/sentry-javascript/pull/14624