feat(node): Do not overwrite prisma `db.system` in newer Prisma versions (#14771)
Since v5.22, Prisma properly adds the `db.system` to it's OTEL spans. 
This PR fixes the behavior so we do not overwrite it. For now, I left
the fallback behavior in, that would add it if `db.system` is missing -
as otherwise this would "break" on older prisma versions.

This also needs to be backported to v8.