feat!: Require Node `>=18` as minimum supported version (#14749)
We decided to require node 18+ in v9. This is to align with the upcoming
changes in OpenTelemetry. We also believe it is a reasonable version
requirement - Node 16 is EOL for some time now. This will also allow us
to update a bunch of dev tooling (vitest, ...) properly and streamline
some processes.

For users requiring Node 14/16, they can remain on v8 (which will be
continued to be supported with bug fixes for some time).

ref https://github.com/getsentry/sentry-javascript/issues/14257