feat(core)!: Remove `getDomElement` method (#14797)
I noticed this in
https://github.com/getsentry/sentry-javascript/pull/14758, this seems
unnecessary and all browsers/envs we support should be able to handle
this. We only use this in two places anyhow, so we can just guard there
(a bit) for this instead.

I will add a deprecation to v8 for this too, although this is just
exported from core and will most likely not affect anybody.