fix(node): Ensure express requests are properly handled (#14850)
Fixes https://github.com/getsentry/sentry-javascript/issues/14847

After some digging, I figured out that the `req.user` handling on
express is not working anymore, because apparently express does not
mutate the actual http `request` object, but clones/forkes it (??)
somehow. So since we now set the request in the
SentryHttpInstrumentation generally, it would not pick up
express-specific things anymore.

IMHO this is not great and we do not want this anymore in v9 anyhow, but
it was the behavior before. This PR fixes this by setting the express
request again on the isolation scope in an express middleware, which is
registered by `Sentry.setupExpressErrorHandler(app)`.

Note that we plan to change this behavior here:
https://github.com/getsentry/sentry-javascript/pull/14806 but I figured
it still makes sense to fix this on develop first, so we have a proper
history/tests for this. I will backport this to v8 too. Then, in PR
#14806 I will remove the middleware again.