feat: Avoid class fields alltogether
We already have an eslint rule to avoid class fields, but had exceptions for static fields as well as for arrow functions.

This also leads to bundle size increases, so removing the exceptions and handling the (few) exceptions we have there should save some bytes.

Additionally, this has additional challenges if we want to avoid/reduce polyfills, as class fields need to be polyfilled for ES2020, sadly.