feat(core)!: Always use session from isolation scope (#14860)
This PR ensures that we always take the session from the isolation
scope, never from the current scope.

This has the implication that we need to be sure to pass the isolation
scope to `_processEvent`, as this is where the session may be marked as
errored. For this, I updated the internal method `_processEvent` to take
the isolation scope as last argument, as well as streamlining this
slightly.

I opted to update the signature of the protected `_prepareEvent` method
too, and make currentScope/isolationScope required there. We already
always pass this in now, so it safes a few bytes to avoid the fallback
everywhere. This should not really affect users unless they overwrite
the `_processEvent` method, which is internal/private anyhow, so IMHO
this should be fine. I added a small note to the migration guide anyhow!