fix(feedback/v8): Avoid lazy loading code for `syncFeedbackIntegration`
Closes https://github.com/getsentry/sentry-javascript/issues/14891

This PR updates the sync feedback integration to avoid it pulling in the
`lazyLoadIntegration` code. This is not really needed and leads to some
problems (and also bundle size increase).

For this I updated the type signature of `buildFeedbackIntegration` to
ensure that _either_ `lazyLoadIntegration` _or_ the getter functions are
provided, so we can type-safely use this.

We probably also want to backport this to v8 then.