feat(build)!: Drop pre-ES2020 polyfills  (#14882)
This PR updates to use a forked version of sucrase
(https://github.com/getsentry/sucrase/tree/es2020-polyfills).
On this branch, a new option `disableES2019Transforms` is added, which
can be used to only polyfill ES2020 features:

* numeric separators (e.g. `10_000`)
* class fields (e.g. `class X { field; }`)

It also adds a lint step that checks all build output for ES2020
compatibility, to avoid regressions in this regard.