fix(core): Ensure debugIds are applied to all exceptions in an event (#14881)
While investigating
https://github.com/getsentry/sentry-javascript/issues/14841, I noticed
that we had some brittle non-null assertions in our `applyDebugIds`
function which would cause the debug id application logic to terminate
early, in case we'd encounter an `event.exception.values` item without a
stack trace. The added regression test illustrates the scenario in which
debug ids would not have been applied to the second exception prior to
this fix.