feat(solidstart)!: Default to `--import` setup and add `autoInjectServerSentry` (#14862)
This PR adds a `withSentry` wrapper for SolidStart's config to build and
place `instrument.server.ts` alongside the server build output so that
it doesn't have to be placed in `/public` anymore to be discoverable.

The setup is changed to be aligned with Nuxt. 
First, the `instrument.server.ts` file is added to the build output (the
sentry release injection file needs to be copied as well - this is not
ideal at the moment as there **could** be other imports as well, but
it's okay for now)

Then, there are two options to set up the SDK:
1. Users provide an `--import` CLI flag to their start command like
this:
```node --import ./.output/server/instrument.server.mjs .output/server/index.mjs```
2. Users can add `autoInjectServerSentry: 'top-level-import'` and the Sentry config will be imported at the top of the server entry

```typescript
// app.config.ts
import { defineConfig } from '@solidjs/start/config';
import { withSentry } from '@sentry/solidstart';

export default defineConfig(withSentry(
    { /* ... */ },
    { 
      autoInjectServerSentry: 'top-level-import' // optional
    })
 );
```

--- 
builds on top of the idea in this PR: https://github.com/getsentry/sentry-javascript/pull/13784

---------

Co-authored-by: Andrei Borza <andrei.borza@sentry.io>