test(react): Ensure react router 3 tests always have correct types (#14949)
Noticed this failing every now and then locally. After looking into
this, I found this to be a bit weird anyhow:

1. We used react-router-5 types for react-router 4, seems off 🤔 I
adjusted this to v4
2. We used to add the `@types/react-router-v3` alias in package.json,
but this was not being picked up anyhow properly. The reason is that
this version of the types package basically re-exports a bunch of stuff
from a path like `react-router/lib` which our resolution then sometimes
(?) picks up from the v6 react-router package. Which is also why we had
to overwrite this somehow, which is what sometimes failed (?). Now, we
simply define these types in test/globals.d.ts properly, the same way as
before, but there should be no more conflicts and we can safe installing
one unecessary package.