feat(core): Emit client reports for unsampled root spans on span start (#14936)
With this PR, the `sample_rate`,`transaction` client report is now
consistently emitted at the time when the root span is sampled.

Previously, in Node (OTEL) we did not emit this at all, while in browser
we emit it at span end time. This is inconsistent and not ideal.
Emitting it in OTEL at span end time is difficult because `span.end()`
is a no-op there and you do not get access to it anywhere. So doing it
at span start (=sample time) is easier, and also makes more sense as
this is also actually the time when the span is dropped.