ref(browser): Improve active span handling for `browserTracingIntegration` (#14959)
Extracted this out of
https://github.com/getsentry/sentry-javascript/pull/14955

We used to rely on implied stuff here quite a bit, which breaks if we
start returning non recording spans. Honestly this just surfaces that
this is not really ideal as it is 😬 We already pass the client around
there everywhere, so this PR updates this so we keep the active idle
span as non-enumerable prop on the client, ensuring this is consistent
and "pure".