feat(core)!: Update `requestDataIntegration` handling (#14806)
This PR cleans up the behavior of `requestDataIntegration`. For this,
there are a bunch of parts that come together:

1. Removed the `addNormalizedRequestDataToEvent` export, this does not
really do much, you can just directly add the data to the event.
2. Removed the `RequestDataIntegrationOptions` type
3. Stops setting `request` on SDK processing metadata, now only relying
on `normalizedRequest`
4. Streamlined code related to this, taking it out of `utils-hoist`.
Now, all the code related directly to the requestDataIntegration is in
that file, while the more general utils are in core/src/utils/request.ts
5. Also moved the cookie & getClientIpAddress utils out of utils-hoist
6. Added tests for the request utils, we did not have any unit tests
there...
7. Streamlined the header extraction to avoid debug logs and simply
try-catch there normally
8. Streamlined the request extraction to avoid weird urls if no host is
found.

Closes https://github.com/getsentry/sentry-javascript/issues/14297