feat(aws): Rename AWS lambda layer name to `SentryNodeServerlessSDKv9` (#14927)
We decided to stop publishing layers under the `SentryNodeServerlessSDK`
name and instead use the version suffix even for the next major.

This will not break the docs, but will break the product as explained in
https://github.com/getsentry/sentry/issues/82646. I think we can
completely remove that check once we are ready to ship v9 but need to
investigate further, otherwise we follow the approach outlined in that
issue.